cycle.java
Finds the length of one complete cycle of linear congruential pseudo-random number generator. This is done by having 2 iterations done at the same time where one iteration goes through the steps twice as fast as the other, meaning that after every iteration, the two iterations get closer to each other by one step. Once the two iterators hit the same value, this value is stored, and the iteration is repeated at normal speed until the same value is reached so that one full cycle occurs. The length of this cycle is calculated with a counter.